requirements.txt
aiohttp
tensorboard
python-dotenv
PyYAML